+++ lib/hpke_for_testing.ml
@@ -1,0 +1,2 @@
+let setup_base_sender = Hpke.Private.setup_base_sender_with_ephemeral
+let setup_psk_sender = Hpke.Private.setup_psk_sender_with_ephemeral